@@ -15,21 +15,19 @@
         }
     },
     "require": {
-        "php": "^7.2",
+        "php": "^8.0|^8.1",
         "symfony/polyfill-php80": "^1.23"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.0",
+        "phpunit/phpunit": "^9.5",
         "phpstan/phpstan": "^0.12",
         "phpstan/phpstan-phpunit": "^0.12",
-        "overtrue/phplint": "^2.0",
         "chiron/coding-standard": "^3.0"
     },
     "scripts": {
         "phpunit": "phpunit --colors=always",
         "test-coverage": "phpunit --coverage-clover clover.xml",
         "phpstan": "phpstan analyse --ansi",
-        "phplint": "phplint --ansi",
         "check-style": "phpcs src tests",
         "fix-style": "phpcbf src tests"
     }
