@@ -9,23 +9,28 @@ use InvalidArgumentException;
 //https://github.com/illuminate/support/blob/master/Str.php
 //https://github.com/fuelphp/common/blob/master/src/Str.php
 
+// TODO : permettre de chainer les interactions : ex : Str::of('my_string')->explode('\n')->filter()->each($callable)
+// https://github.com/illuminate/support/blob/master/Stringable.php
+// https://github.com/illuminate/support/blob/master/Str.php#L53
+
 final class Str
 {
     /**
      * Starts the $haystack string with the prefix $needle?
      */
+    // TODO : enlever cette méthode et utiliser directement dans le code la méthode native php8 !!!
     public static function startsWith(string $haystack, string $needle): bool
     {
-        return strncmp($haystack, $needle, strlen($needle)) === 0;
+        return str_starts_with($haystack, $needle);
     }
-
 
     /**
      * Ends the $haystack string with the suffix $needle?
      */
+    // TODO : enlever cette méthode et utiliser directement dans le code la méthode native php8 !!!
     public static function endsWith(string $haystack, string $needle): bool
     {
-        return $needle === '' || substr($haystack, -strlen($needle)) === $needle;
+        return str_ends_with($haystack, $needle);
     }
 
     /**
@@ -35,6 +40,7 @@ final class Str
      * @param  string  $search
      * @return string
      */
+    // TODO : renommer $subject en $haystack et $search par $needle ???
     public static function before(string $subject, string $search): string
     {
         if ($search === '') {
@@ -53,8 +59,10 @@ final class Str
      * @param  string  $search
      * @return string
      */
+    // TODO : renommer $subject en $haystack et $search par $needle ???
     public static function after(string $subject, string $search): string
     {
+        // TODO - Autre exemple : https://github.com/nette/utils/blob/a828903f85bb513e51ba664b44b61f20d812cf20/src/Utils/Strings.php#L434
         return $search === '' ? $subject : array_reverse(explode($search, $subject, 2))[0];
     }
 
@@ -68,6 +76,7 @@ final class Str
      */
     public static function substr(string $string, int $start, ?int $length = null): string
     {
+        // TODO : il faut ajouter dans le fichier composer que l'extension mb_string est obligatoire !!!!
         return mb_substr($string, $start, $length, 'UTF-8');
     }
 
@@ -79,6 +88,7 @@ final class Str
      * @param  bool  $ignoreCase
      * @return bool
      */
+    // TODO : utiliser plutot cette méthode, voir même la méthode native de php8 ????  https://github.com/nette/utils/blob/a828903f85bb513e51ba664b44b61f20d812cf20/src/Utils/Strings.php#L97
     public static function contains($haystack, $needles, $ignoreCase = false)
     {
         if ($ignoreCase) {
